Include 'disease' field on ResProbForecast
